backend/requirements.txt
Flask
flask_cors
mysql-connector-python
